feat: Esse commit cria classes para os endpoints já criados!
