+++ template-labook/src/dataBase/migrations.ts
@@ -1,4 +1,4 @@
-import { connection } from "./index"
+import { connection } from "./connection"
 
 connection
    .raw(`
